ICC cleanup for IA32 & x64
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/branches/UefiHiiEnable4354@4567 6f19259b-4bc3-4df7-8a09-765794883524

+++ MdeModulePkg/Universal/UefiHiiDatabaseDxe/Image.c
@@ -374,7 +374,7 @@ Output4bitPixel (
   Palette = AllocateZeroPool (PaletteSize);
   ASSERT (Palette != NULL);
   CopyMem (Palette, PaletteInfo, PaletteSize);
-  PaletteNum = Palette->PaletteSize / sizeof (EFI_HII_RGB_PIXEL);
+  PaletteNum = (UINT16)(Palette->PaletteSize / sizeof (EFI_HII_RGB_PIXEL));
 
   ZeroMem (PaletteValue, sizeof (PaletteValue));
   CopyRgbToGopPixel (PaletteValue, Palette->PaletteValue, PaletteNum);
@@ -446,7 +446,7 @@ Output8bitPixel (
   Palette = AllocateZeroPool (PaletteSize);
   ASSERT (Palette != NULL);
   CopyMem (Palette, PaletteInfo, PaletteSize);
-  PaletteNum = Palette->PaletteSize / sizeof (EFI_HII_RGB_PIXEL);
+  PaletteNum = (UINT16)(Palette->PaletteSize / sizeof (EFI_HII_RGB_PIXEL));
   ZeroMem (PaletteValue, sizeof (PaletteValue));
   CopyRgbToGopPixel (PaletteValue, Palette->PaletteValue, PaletteNum);
   SafeFreePool (Palette);
